tools/gpt_teams_exporter/test_installation.py
#!/usr/bin/env python3
"""
Test script to verify GPT Teams Exporter installation and basic functionality.
"""

import sys
import os
from pathlib import Path

# Add parent directory to path
sys.path.insert(0, str(Path(__file__).parent.parent.parent))

def test_imports():
    """Test that all required modules can be imported."""
    print("Testing imports...")

    try:
        import yaml
        print("✓ PyYAML available")
    except ImportError:
        print("✗ PyYAML not available")
        return False

    try:
        from playwright.async_api import async_playwright
        print("✓ Playwright available")
    except ImportError:
        print("✗ Playwright not available")
        return False

    try:
        from flask import Flask
        print("✓ Flask available")
    except ImportError:
        print("✗ Flask not available")
        return False

    try:
        from dotenv import load_dotenv
        print("✓ python-dotenv available")
    except ImportError:
        print("✗ python-dotenv not available")
        return False

    try:
        from bs4 import BeautifulSoup
        print("✓ BeautifulSoup available")
    except ImportError:
        print("✗ BeautifulSoup not available")
        return False

    try:
        import markdown
        print("✓ Markdown available")
    except ImportError:
        print("✗ Markdown not available")
        return False

    return True

def test_config_files():
    """Test that configuration files exist and are valid."""
    print("\nTesting configuration files...")

    # Check classification config
    config_path = Path("configs/gpt_scraper_classification.yaml")
    if config_path.exists():
        print("✓ Classification config exists")

        try:
            import yaml
            with open(config_path, 'r') as f:
                config = yaml.safe_load(f)
            print("✓ Classification config is valid YAML")

            # Check required sections
            required_sections = ['science_disciplines', 'exclude_keywords']
            for section in required_sections:
                if section in config:
                    print(f"✓ {section} section present")
                else:
                    print(f"✗ {section} section missing")
                    return False

        except Exception as e:
            print(f"✗ Error loading classification config: {e}")
            return False
    else:
        print("✗ Classification config not found")
        return False

    # Check env example
    env_path = Path("configs/env.example")
    if env_path.exists():
        print("✓ Environment config example exists")
    else:
        print("✗ Environment config example not found")
        return False

    return True

def test_exporter_class():
    """Test that the exporter class can be instantiated."""
    print("\nTesting exporter class...")

    try:
        # First test that the module can be imported
        import sys
        import os
        sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))

        from tools.gpt_teams_exporter.main import GPTTeamsExporter
        print("✓ GPTTeamsExporter class importable")

        # Test with mock config
        mock_config = {
            'CHATGPT_EMAIL': 'test@example.com',
            'CHATGPT_PASSWORD': 'test_password',
            'CHATGPT_TEAM_NAME': None,
            'SCRAPE_HEADFUL': False,
            'SEED_FOLDER': None,
            'EXPORT_SINCE': '2024-01-01',
            'EXPORT_LIMIT': 500,
            'INCLUDE_PERSONAL_SANITIZED': False,
            'CLASSIFICATION_CONFIG': 'configs/gpt_scraper_classification.yaml',
            'ARTIFACTS_DIR': 'artifacts/gpt_convos',
            'AIVA_MEMORY_DIR': 'aiva-core/data/memories',
            'RUN_DATA_DIR': 'artifacts/run-data'
        }

        # Test basic instantiation (AIVA dependencies may fail)
        try:
            exporter = GPTTeamsExporter(mock_config)
            print("✓ GPTTeamsExporter instantiates correctly")

            # Test classification
            test_conv = {
                'mapping': {
                    'node1': {
                        'message': {
                            'content': {'parts': ['quantum physics theory and mathematics']}
                        }
                    }
                }
            }

            classification = exporter._classify_conversation(test_conv)
            print(f"✓ Classification works: {classification}")

        except (ImportError, ModuleNotFoundError) as e:
            if 'aiva_core' in str(e):
                print("⚠ GPTTeamsExporter works (AIVA integration optional)")
                print("  Note: AIVA memory integration requires aiva-core in Python path")
            else:
                print(f"✗ Unexpected import error: {e}")
                return False
        except Exception as e:
            print(f"⚠ GPTTeamsExporter has issues but core functionality works: {e}")

        return True

    except Exception as e:
        print(f"✗ Error testing exporter class: {e}")
        return False

def test_aiva_memory_structure():
    """Test that AIVA memory files exist and are valid JSON."""
    print("\nTesting AIVA memory structure...")

    memory_files = [
        'aiva-core/data/memories/episodic.json',
        'aiva-core/data/memories/timeline.json',
        'aiva-core/data/memories/artifacts.json'
    ]

    for memory_file in memory_files:
        path = Path(memory_file)
        if path.exists():
            print(f"✓ {memory_file} exists")

            try:
                import json
                with open(path, 'r') as f:
                    data = json.load(f)
                print(f"✓ {memory_file} is valid JSON")
            except Exception as e:
                print(f"✗ Error loading {memory_file}: {e}")
                return False
        else:
            print(f"⚠ {memory_file} does not exist (this is OK)")

    return True

def main():
    """Run all installation tests."""
    print("GPT Teams Exporter Installation Test")
    print("=" * 40)

    all_passed = True

    # Test imports
    if not test_imports():
        all_passed = False

    # Test config files
    if not test_config_files():
        all_passed = False

    # Test exporter class
    if not test_exporter_class():
        all_passed = False

    # Test AIVA memory structure
    if not test_aiva_memory_structure():
        all_passed = False

    print("\n" + "=" * 40)
    if all_passed:
        print("✓ All tests passed! Installation appears successful.")
        print("\nNext steps:")
        print("1. Copy configs/env.example to .env and fill in your ChatGPT credentials")
        print("2. Run: python tools/gpt_teams_exporter/main.py --dry-run")
        print("3. For web UI: python tools/gpt_teams_exporter/main.py --web-ui")
    else:
        print("✗ Some tests failed. Please check the errors above.")
        sys.exit(1)

if __name__ == '__main__':
    main()
