requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
requests>=2.25.0
scikit-learn>=1.0.0
tensorflow>=2.8.0  # Optional, for comparison
torch>=1.11.0     # Optional, for comparison
jupyter>=1.0.0    # Optional, for notebooks
seaborn>=0.11.0   # Optional, for visualization
playwright>=1.40.0
flask>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.4.0